# docs/swagger.yaml
basePath: /api
definitions:
  controller.CardContentReq:
    properties:
      text:
        type: string
    required:
    - text
    type: object
...
  controller.CollapseDecksRequest:
    properties:
      source_deck_id:
        type: string
      target_deck_id:
        type: string
    required:
    - source_deck_id
    - target_deck_id
    type: object
  controller.CreateCardRequest:
    properties:
      back:
        $ref: '#/definitions/controller.CardContentReq'
      deck_id:
        type: string
      front:
        $ref: '#/definitions/controller.CardContentReq'
      link:
        type: string
    required:
    - back
    - deck_id
    - front
    type: object
...
  controller.UpdateCardRequest:
    properties:
      back:
        $ref: '#/definitions/controller.CardContentReq'
      front:
        $ref: '#/definitions/controller.CardContentReq'
      link:
        type: string
    type: object
  controller.UpdateDeckRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  echo.HTTPError:
    properties:
      message: {}
    type: object
  types.Card:
    properties:
      back:
        $ref: '#/definitions/types.CardBack'
      created_at:
        type: string
      deck_id:
        type: string
      fail_count:
        type: integer
      front:
        $ref: '#/definitions/types.CardFront'
      id:
        type: string
      link:
        type: string
      pass_count:
        type: integer
      retired:
        type: boolean
      reviewed_at:
        type: string
      skip_count:
        type: integer
      star_rating:
        type: integer
      updated_at:
        type: string
    type: object
  types.CardBack:
    properties:
      text:
        type: string
    type: object
  types.CardFront:
    properties:
      text:
        type: string
    type: object
...
  types.Deck:
    properties:
      cards:
        items:
          $ref: '#/definitions/types.Card'
        type: array
      description:
        description: New Description field
        type: string
      id:
        description: UUID string as the primary key
        type: string
      name:
        type: string
    type: object
host: 192.168.86.176:8789
info:
  contact: {}
  description: API documentation for the MeowMorize Flashcard App.
  title: MeowMorize Flashcard API
  version: "1.0"
paths:
  /cards:
    post:
      consumes:
      - application/json
      description: Create a new card and associate it with a deck
      parameters:
      - description: Create Card
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/controller.CreateCardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a new card
      tags:
      - cards
  /cards/{id}:
    delete:
      description: Delete a card by its ID
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a card
      tags:
      - Cards
    get:
      description: Retrieve a single card by its ID
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Card'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a card by ID
      tags:
      - Cards
    put:
      consumes:
      - application/json
      description: Update the details of an existing card by its ID
      parameters:
      - description: Card ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Card
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateCardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Card'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Update an existing card
      tags:
      - cards
...
  /decks:
    get:
      description: Retrieve a list of all decks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.Deck'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all decks
      tags:
      - Decks
    post:
      consumes:
      - application/json
      description: Create a new deck with provided details
      parameters:
      - description: Deck to create
        in: body
        name: deck
        required: true
        schema:
          $ref: '#/definitions/types.Deck'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/types.Deck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new deck
      tags:
      - Decks
  /decks/{id}:
    delete:
      description: Delete a deck by its ID
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a deck
      tags:
      - Decks
    get:
      description: Retrieve a single deck by its ID
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Deck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a deck by ID
      tags:
      - Decks
    put:
      consumes:
      - application/json
      description: Update the name (and optionally the ID) of an existing deck by
        its ID
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Deck
        in: body
        name: deck
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateDeckRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.Deck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a deck
      tags:
      - Decks
  /decks/collapse:
    post:
      consumes:
      - application/json
      description: Merge all cards from the source deck into the target deck by removing
        each card from the source deck, deleting it, and adding it to the target deck.
      parameters:
      - description: Deck IDs to collapse
        in: body
        name: collapse
        required: true
        schema:
          $ref: '#/definitions/controller.CollapseDecksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Collapse two decks
      tags:
      - Decks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"